LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/script-access.tentative-expected.txt

CONSOLE MESSAGE: TypeError: template.getInnerHTML is not a function. (In 'template.getInnerHTML({includeShadowRoots: true})', 'template.getInnerHTML' is undefined)

Harness Error (FAIL), message = TypeError: template.getInnerHTML is not a function. (In 'template.getInnerHTML({includeShadowRoots: true})', 'template.getInnerHTML' is undefined)
